Remove dependence on special type checking for int/nat (#192)
* To prepare for removing custom type-checking support for int/nat,
have the verus macro transform spec arithmetic expressions into calls to
various Spec* traits and spec_* functions.

* Turn on --no-enhanced-typecheck in all tests

* Turn on --no-enhanced-typecheck in regression.rs

* Fix some missing :: in ::builtin, and remove unused code

* Enable SpecEuclideanDiv and SpecEuclideanMod traits

* cargo fmt

* Have SpecEuclideanDiv return Self for unsigned, int for signed

* Add clipping for is_spec_arith_binary

* For nat + unsigned and nat * unsigned, return nat instead of int

* Add spec traits for binary bitwise operations

+++ source/pervasive/state_machine_internal.rs
@@ -117,7 +117,7 @@ pub fn assert_general_guard_map(b: bool) { requires(b); ensures(b); }
 impl<A> Seq<A> {
     #[spec] #[verifier(publish)]
     pub fn update_at_index(self, i: int, a: A) -> Self {
-        recommends(0 <= i && i < self.len());
+        recommends(0 <= i && i < self.len() as int);
 
         self.update(i, a)
     }
